Style: 채팅방 이름 변경 모달 컴포넌트 분리
